requirements.txt
beautifulsoup4==4.11.1
discord.py==1.7.3
pydantic==1.10.7
python_dateutil==2.8.2
Requests==2.31.0
